As after '3.3 Home Page', with changes.
Changes needed because the [tutorial](https://www.sigient.com/blog/movie-listings-application-with-react-router-v-4) which this repo walks through **does not** use the actually-released React Router v4, but instead documents a (fairly early) prerelease, and so numerous changes occurred between that time and this. This will confuse the heck out of a novice to React who's simply trying to wrap her head around routing.