Add full Movie element tests using dummy data
This shows how *creaky* the all-singing, all-dancing Movie component is as it
presently stands. Were this component destined for production, even as a side
project, we'd quickly decompose it into multiple child components to reduce the
excessively brittle monolithic tests which don't (and can't) test multiple
scenarios. What if a Movie has no reviews? Isn't the "presentation" of the Movie
metadata *embarrassing?* If we *do* ever change this to use a design language
(see the [README](https://github.com/jdickey/flix/tree/master#design-language-or-css-framework)),
any of the tests that mention CSS styling will break, because they assume no other
styling is being done. And so it goes. 😪