Use MovieData rather than JSON for Home component
One less component that cares about JSON data.

Incidentally, we've changed the value of the 'key' React uses for managing list items. It's highly unlikely that two movies with the same title and same director would be in the Top Four list; that combination works just fine for a key value that is independent of list ordering.