Use data from "remote service" -- except when we can't
We had this all sewn up some time ago, after success with the  third fetch-data-from-a-URL library that we tried,  and then we remembered: our CI executes in a single pipeline (thread of execution). This means, among other things, that we can't make requests to a background server that we can't _start_ in the background. Oops.

Our MovieData class still provides the façade for our data, and it still gets that data from the remote service _when it's not running with the `CI` environment variable set._ When it is, it falls back to local data — which, since the JSON file has been updated to reflect `json-server`'s requirements, requires a slight massage.

For those who think, "hey, isn't this feature gallop", we refer you to the history of Knuth's TAOCP.