Abstract out movie data (#7)
* Update dependencies

* Add MovieData (non-component) object

Object has two functions, `find` and `get`, as specified in Issue #6. Currently uses hard-coded JSON file; this will change soon.

Also fixed typo in JSON data file from original tutorial, which was stumbled on by chance. Pffffft.

* Freeze object containing read-only data

* Adapt Movie to new MovieData data source

Movie no linger knows or cares that the data is coming (for now) from a JSON file.

* Use MovieData rather than JSON for Home component

One less component that cares about JSON data.

Incidentally, we've changed the value of the 'key' React uses for managing list items. It's highly unlikely that two movies with the same title and same director would be in the Top Four list; that combination works just fine for a key value that is independent of list ordering.

* Fully compartmentalise JSON seed data in MovieData

* Simplify test and get coverage back to 100%

Win! :sunglasses:

* Use data from "remote service" -- except when we can't

We had this all sewn up some time ago, after success with the  third fetch-data-from-a-URL library that we tried,  and then we remembered: our CI executes in a single pipeline (thread of execution). This means, among other things, that we can't make requests to a background server that we can't _start_ in the background. Oops.

Our MovieData class still provides the façade for our data, and it still gets that data from the remote service _when it's not running with the `CI` environment variable set._ When it is, it falls back to local data — which, since the JSON file has been updated to reflect `json-server`'s requirements, requires a slight massage.

For those who think, "hey, isn't this feature gallop", we refer you to the history of Knuth's TAOCP.

* Fix ESLint errors from Jest names "not being defined"

We had the right plugin installed; we just forgot to flip the switch in `.eslintrc.js` to _use_ it. Oops. : disappointed_relieved:

* Update README re data source and Docker [ci skip]