Quickly decompose Movie into child components
First whack at breaking the massive, brittle Movie component into smaller pieces that can each be extracted, covered with bespoke tests, and then progressively redone as a tree of related components (e.g., Movie.Title, Movie.Reviews, etc).

Once we have that, implementing a parallel design-language-aware component tree (e.g., SUIMovie.Title, SUIMovie.Reviews) is extremely straightforward.

This isn't the _fastest_ way to get to where we want to be if _every bit of work is absolutely perfect,_ but the next time that happens on schedule, to expectations, will be the very first.